Break book tests out into sub modules
